refactor(ui,iso): rename use-render.ts to render-element.ts
The file exported renderElement (a pure function) but kept the vestigial
use- prefix from when it was useRender. Rename the files in both packages
to match; pure mechanical move, no logic change.
