+++ packages/ui/src/dialog/dialog.tsx
@@ -6,7 +6,7 @@ import {
   useMemo,
   useRef,
 } from 'preact/hooks';
-import { renderElement, type RenderProp } from '../use-render.js';
+import { renderElement, type RenderProp } from '../render-element.js';
 import { useControllableState } from '../use-controllable-state.js';
 import { useDescriptionRegistry } from '../use-description-registry.js';
 import { mergeRefs } from '../merge-refs.js';
