+++ packages/ui/src/index.ts
@@ -1,7 +1,7 @@
 // Public barrel for @hono-preact/ui. Primitives and components are exported
 // here as they land in subsequent tasks.
 export { mergeRefs } from './merge-refs.js';
-export { renderElement, type RenderProp } from './use-render.js';
+export { renderElement, type RenderProp } from './render-element.js';
 export { useControllableState } from './use-controllable-state.js';
 export {
   usePosition,
