+++ packages/ui/src/menu/menu.tsx
@@ -7,7 +7,7 @@ import {
   type VNode,
 } from 'preact';
 import { useContext, useId, useLayoutEffect, useMemo } from 'preact/hooks';
-import { renderElement, type RenderProp } from '../use-render.js';
+import { renderElement, type RenderProp } from '../render-element.js';
 import { useControllableState } from '../use-controllable-state.js';
 import type { Side, Align } from '../use-position.js';
 import { useMenuCore } from './use-menu-core.js';
