+++ packages/ui/src/positioner.tsx
@@ -1,6 +1,6 @@
 import { h, type ComponentChildren, type JSX, type RefObject } from 'preact';
 import { useMemo } from 'preact/hooks';
-import { renderElement, type RenderProp } from './use-render.js';
+import { renderElement, type RenderProp } from './render-element.js';
 import { usePositioner } from './use-positioner.js';
 import type { Side, Align, ClientRectGetter } from './use-position.js';
 import { PositionerContext } from './positioner-context.js';
